ChatGPT API/requirements.txt
json
python-dot_env
openai